+++ eventmesh-openconnect/eventmesh-openconnect-java/src/main/java/org/apache/eventmesh/openconnect/api/config/Config.java
@@ -15,17 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.api.config;
+package org.apache.eventmesh.openconnect.api.config;
 
-public class Constants {
-
-    public static final String ENV_TARGET = "connectorTarget";
-
-    public static final String ENV_PORT = "connectorPort";
-
-    public static final String ENV_CONFIG_FILE = "connectorConf";
-
-    public static final int DEFAULT_ATTEMPT = 3;
-
-    public static final int DEFAULT_PORT = 8080;
+public abstract class Config {
 }
