+++ eventmesh-openconnect/eventmesh-openconnect-java/src/main/java/org/apache/eventmesh/openconnect/api/data/DefaultKeyValue.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.api.data;
+package org.apache.eventmesh.openconnect.api.data;
 
 import java.util.Map;
 import java.util.Set;
