+++ eventmesh-openconnect/eventmesh-openconnect-java/src/main/java/org/apache/eventmesh/openconnect/api/data/RecordPosition.java
@@ -15,27 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.api.data;
+package org.apache.eventmesh.openconnect.api.data;
 
-import java.util.Map;
 import java.util.Objects;
 
-public class RecordPartition {
+public class RecordPosition {
 
-    /**
-     * if pull message from mq key=topic,
-     * value=topicName key=brokerName,
-     * value=brokerName key=queueId,
-     * value=queueId
-     */
-    private final Map<String, ?> partition;
+    private final RecordPartition recordPartition;
 
-    public RecordPartition(Map<String, ?> partition) {
-        this.partition = partition;
+    private final RecordOffset recordOffset;
+
+    public RecordPosition(RecordPartition recordPartition, RecordOffset recordOffset) {
+        this.recordPartition = recordPartition;
+        this.recordOffset = recordOffset;
     }
 
-    public Map<String, ?> getPartition() {
-        return partition;
+    public RecordPartition getPartition() {
+        return recordPartition;
+    }
+
+    public RecordOffset getOffset() {
+        return recordOffset;
     }
 
     @Override
@@ -43,22 +43,15 @@ public class RecordPartition {
         if (this == o) {
             return true;
         }
-        if (!(o instanceof RecordPartition)) {
+        if (!(o instanceof RecordPosition)) {
             return false;
         }
-        RecordPartition partition1 = (RecordPartition) o;
-        return Objects.equals(partition, partition1.partition);
+        RecordPosition position = (RecordPosition) o;
+        return recordPartition.equals(position.recordPartition) && recordOffset.equals(position.recordOffset);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(partition);
-    }
-
-    @Override
-    public String toString() {
-        return "RecordPartition{"
-            + "partition=" + partition
-            + "}";
+        return Objects.hash(recordPartition, recordOffset);
     }
 }
